Rename and move to package
