+++ src/jvmTest/kotlin/choliver/nespot/playtest/GamesTest.kt
@@ -1,0 +1,41 @@
+package choliver.nespot.playtest
+
+import choliver.nespot.playtest.engine.Engine
+import choliver.nespot.playtest.engine.Engine.Mode.COMPARE
+import org.junit.jupiter.api.Test
+import java.io.File
+
+class GamesTest {
+  @Test
+  fun `super mario bros`() = engine.execute("smb")
+
+  @Test
+  fun `super mario bros 3`() = engine.execute("smb3")
+
+  @Test
+  fun `bubble bobble`() = engine.execute("bb")
+
+  @Test
+  fun `donkey kong`() = engine.execute("dk")
+
+  @Test
+  fun `kirby's adventure`() = engine.execute("kirby")
+
+  @Test
+  fun `micro machines`() = engine.execute("mm")
+
+  @Test
+  fun `mig-29 soviet fighter`() = engine.execute("mig29")
+
+  @Test
+  fun castelian() = engine.execute("castelian")
+
+  @Test
+  fun `jurassic park`() = engine.execute("jp")
+
+  private val engine = Engine(
+    romsBase = File("roms"),
+    capturesBase = File("captures"),
+    mode = COMPARE
+  )
+}
