+++ src/jvmTest/kotlin/choliver/nespot/playtest/engine/CapturingVideoSink.kt
@@ -1,4 +1,4 @@
-package choliver.nespot.playtest
+package choliver.nespot.playtest.engine
 
 import choliver.nespot.SCREEN_HEIGHT
 import choliver.nespot.SCREEN_WIDTH
